refactor: 패키지별 분류
