+++ src/main/java/ladder/dto/InputManagement.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.dto;
 
 import java.util.List;
 import java.util.stream.Collectors;
