+++ src/test/java/ladder/NameTest.java
@@ -1,5 +1,6 @@
 package ladder;
 
+import ladder.dto.Name;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
